move spread functions to own file
