+++ src/store/blocks/api.ts
@@ -1,34 +1,8 @@
 import { Block, BlockSaveBase, MetricsResult, BlockBase } from './state';
 
 import { get, put, patch } from '@/core/fetch';
+import { spreadData, unspreadData } from '@/core/api-spread';
 
-export function spreadData(input: any) {
-  const spreadInput = {
-    ...input.data,
-    ...input,
-  };
-
-  delete spreadInput.data;
-
-  return spreadInput;
-}
-
-export function unspreadData(input: any) {
-  const { id, type } = input;
-  const data = { ...input };
-
-  // remove id and type from data
-  delete data.id;
-  delete data.type;
-
-  return {
-    // conditionally spread id and type
-    ...(id ? { id } : {}),
-    ...(type ? { type } : {}),
-    // include the data object
-    data,
-  };
-}
 
 export function fetchBlock(id: string): Promise<Block> {
   return get(`/blocks/${encodeURIComponent(id)}`).then(block => spreadData(block));
